# back-end/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "TB Tracker API",
    "description": "API pour gérer les patients atteints de la tuberculose",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/patients": {
      "post": {
        "summary": "Ajouter un patient",
        "parameters": [
          {
            "in": "body",
            "name": "patient",
            "description": "Les informations du patient",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "IDPatient": { "type": "integer" },
                "nom": { "type": "string" },
                "prenom": { "type": "string" },
                "email": { "type": "string", "format": "email" },
                "age" :{"type":"integer"},
                "adresse": { "type": "string" },
                "numero": { "type": "string" },
                "sexe": { "type": "string" },
                "DateNaissance": { "type": "string", "format": "date" },
                "poidsInitial": { "type": "number" },
                "categorie": { "type": "integer" },
                "preuve": { "type": "string" },
                "LocalisationTB": { "type": "string" },
                "typeTuberculose": { "type": "string" },
                "comptage_tuberculeux": { "type": "string" },
                "antecedents": { "type": "string" },
                "note": { "type": "string" }
              },
              "required": [
                "IDPatient", "nom", "prenom", "DateNaissance", "poidsInitial",
                "numero", "email", "adresse", "LocalisationTB"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Patient ajouté avec succès"
          },
          "400": {
            "description": "Requête invalide (champs manquants ou mal formatés)"
          },
          "409": {
            "description": "Patient avec cet email existe déjà"
          }
        }
      },
      "get": {
        "summary": "Récupérer la liste des patients",
        "responses": {
          "200": {
            "description": "Liste des patients récupérée avec succès"
          }
        }
      }
    },
    "/patients/{IDPatient}": {
      "get": {
        "summary": "Obtenir un patient spécifique",
        "parameters": [
          {
            "name": "IDPatient",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Détails du patient"
          },
          "404": {
            "description": "Patient non trouvé"
          }
        }
      },
      "put": {
        "summary": "Modifier un patient",
        "parameters": [
          {
            "name": "IDPatient",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "patient",
            "description": "Mise à jour des informations du patient",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nom": { "type": "string" },
                "prenom": { "type": "string" },
                "email": { "type": "string", "format": "email" },
                "age" :{"type":"integer"},
                "adresse": { "type": "string" },
                "numero": { "type": "string" },
                "sexe": { "type": "string" },
                "DateNaissance": { "type": "string", "format": "date" },
                "poidsInitial": { "type": "number" },
                "categorie": { "type": "integer" },
                "preuve": { "type": "string" },
                "LocalisationTB": { "type": "string" },
                "typeTuberculose": { "type": "string" },
                "comptage_tuberculeux": { "type": "string" },
                "antecedents": { "type": "string" },
                "note": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patient mis à jour"
          },
          "404": {
            "description": "Patient non trouvé"
          }
        }
      },
      "delete": {
        "summary": "Supprimer un patient",
        "parameters": [
          {
            "name": "IDPatient",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Patient supprimé"
          },
          "404": {
            "description": "Patient non trouvé"
          }
        }
      }
    }
  }
}